requirements.txt
Django==1.11.29
django-tenant-schemas==1.9.0
six==1.16.0